* Implemented "Transformer" and Parameters for "synchronous" request (request with result, through a callback). * TODO (#1) : Develop transformers for base64_encoded data (user, inventory) and create Classes (?) for Inventory and User.
TO DO:
* Timing to try to reconnect to the websocket if disconnected (X times during Y seconds/minutes).
* Implement an event handler for device events? (just like in Java/.NET)
* Split the API in many files? Using RequireJS ?

+++ src/com/spacecode/smartserver/command/CommandRemoveUser.java
@@ -27,6 +27,12 @@ public class CommandRemoveUser extends ClientCommand
             throw new ClientCommandException("Invalid number of parameters.");
         }
 
+        if(DeviceHandler.getDevice() == null)
+        {
+            SmartServer.sendMessage(ctx, RequestCode.REMOVE_USER, FALSE);
+            return;
+        }
+
         String username = parameters[0];
 
         if(!DeviceHandler.getDevice().getUsersService().removeUser(username))
